Update README.md
@JuliaRegistrator register branch=main

Release notes:

Initial version of the package LightLattices.
The package provides a convenient interface to work with Lattices with arbitrary unit cells.

For the exported types of unit cells and lattices, it defines the array interface which allows to access the coordinate of specific node by its index. In addition to that, it provides a function relative_coordinate which allows to calculate the shortest vector connecting the two nodes.